Rename "app" module to "sample"
