+++ sample/proguard-rules.pro
