+++ sample/src/main/res/font/iran_sans_regular.ttf
