+++ sample/src/main/res/mipmap-mdpi/ic_launcher.png
