+++ sample/src/main/res/values/styles.xml
